Functions

Functions exported from VisAna:

PyPlot.contour — Function

contour(data,var, levels=0; plotrange, plotinterval, kwargs)

Wrapper over the contour function in matplotlib.

PyPlot.contourf — Function

contourf(data, var, levels=0; plotrange, plotinterval, kwargs)

Wrapper over the contourf function in matplotlib.

PyPlot.plot — Method

plot(data, var; kwargs)

Wrapper over the plot function in matplotlib.

PyPlot.plot_surface — Method

plot_surface(data, var; plotrange, plotinterval, kwargs)

Wrapper over the plot_surface function in matplotlib.

PyPlot.plot_trisurf — Method

plot_trisurf(data::Data, var::String;
	plotrange::Vector{Float64}=[-Inf,Inf,-Inf,Inf], kwargs::Dict=Dict())

Wrapper over the plot_trisurf function in matplotlib.

PyPlot.scatter — Method

scatter(data, var; kwargs)

Wrapper over the scatter function in matplotlib.

PyPlot.streamplot — Method

streamplot(data, var; plotrange, plotinterval=0.1, density=1.0, color="")

Wrapper over the streamplot function in matplotlib. Streamplot does not have **kwargs in the API.

PyPlot.tricontourf — Method

tricontourf(data, var; plotrange, plotinterval, kwargs)

Wrapper over the tricontourf function in matplotlib.

VisAna.animatedata — Method

animatedata()

Generate animations from data. This is basically calling plotdata function for multiple snapshots. The main issue here is to determine the colorbar/axis range in advance to avoid any jump in the movie.

VisAna.cutplot — Method

cutplot(data, var; plotrange=[-Inf,Inf,-Inf,Inf], cut=' ',
	plotinterval=0.1, density=1.0, cutPlaneIndex=1,level=20)

2D plane cut contourf of 3D box data.

VisAna.plotdata — Method

plotdata(data, func, (...))

Plot the variable from SWMF output.

plotdata(data, "p", plotmode="contbar")

plotdata(data, "p", plotmode="grid")

plotdata(data, func, plotmode="trimesh",plotrange=plotrange, plotinterval=0.2)

Input arguments

data::Data: output data.

func::String: variables for plotting.

plotmode::String: (optional) type of plotting ["cont","contbar"]...

plotrange::Vector: (optional) range of plotting.

plotinterval: (optional) interval for interpolation.

level: (optional) level of contour.

density: (optional) density for streamlines.

cut: (optional) select 2D cut plane from 3D outputs ["x","y","z"].

cutPlaneIndex: (optional)

multifigure: (optional) 1 for multifigure display, 0 for subplots.

verbose: (optional) display additional information.

Right now this can only deal with 2D plots or 3D cuts. Full 3D plots may be supported in the future.

VisAna.plotlogdata — Method

plotlogdata(data, head, func, plotmode="line")

Plot information from log file.

Input arguments

data::Array: output data.

head::NamedTuple: header info.

func::String: variables for plotting.

plotmode::String: type of plotting ["line","scatter"].

VisAna.streamslice — Method

streamslice(data::Data, var::String;
  plotrange=[-Inf,Inf,-Inf,Inf], cut=' ',
  plotinterval=0.1, density=1.0, cutPlaneIndex=1,color="w")

Plot streamlines on 2D slices of 3D box data. Variable string must be separated with ;. Tranposes aree needed because of meshgrid and ndgrid conversion.