Public APIs

SpaceAnalysis.MVA — Method

MVA(Bx, By, Bz; verbose=false)

Perform minimum variance analysis to vector components defined in orthogonal coordinates Bx, By and Bz. If λ₁ ≥ λ₂ ≥ λ₃ are 3 eigenvalues of the constructed matrix M, then a good indicator of nice fitting LMN coordinate system should have λ₂/λ₃ > 5. Set verbose=true to turn on the check.

SpaceAnalysis.mag2db — Method

Convert x from magnitude to decibels.

SpaceAnalysis.sma — Function

sma(x, n=(5,5))

Simple moving average of x with n[1] leading and n[2] trailing points. Default boundary is "replicate", meaning that the border pixels extend beyond the boundaries.

SpaceAnalysis.sma — Function

sma(x, n=5)

Simple moving box average of the vector data x with box length 'n'. One-sided average on the left and right edge with replicate border.

SpaceAnalysis.spectrum — Method

spectrum(x, Fs)

Return the frequency and amplitude for the single-sided spectrum of vector x given sample frequency of Fs.