Functions

Functions exported from VisAna:

PyPlot.contour — Function

contour(data,var, levels=0; plotrange, plotinterval, kwargs)

Wrapper over the contour function in matplotlib.

PyPlot.contourf — Function

contourf(data, var, levels=0; plotrange, plotinterval, kwargs)

Wrapper over the contourf function in matplotlib.

PyPlot.plot — Method

plot(data, var; kwargs)

Wrapper over the plot function in matplotlib.

PyPlot.plot_surface — Method

plot_surface(data, var; plotrange, plotinterval, kwargs)

Wrapper over the plot_surface function in matplotlib.

PyPlot.plot_trisurf — Method

plot_trisurf(data::Data, var::String;
	plotrange::Vector{Float64}=[-Inf,Inf,-Inf,Inf], kwargs::Dict=Dict())

Wrapper over the plot_trisurf function in matplotlib.

PyPlot.scatter — Method

scatter(data, var; kwargs)

Wrapper over the scatter function in matplotlib.

PyPlot.streamplot — Method

streamplot(data, var; plotrange, plotinterval=0.1, density=1.0, color="")

Wrapper over the streamplot function in matplotlib. Streamplot does not have **kwargs in the API.

PyPlot.tricontourf — Method

tricontourf(data, var; plotrange, plotinterval, kwargs)

Wrapper over the tricontourf function in matplotlib.

VisAna.animatedata — Method

animatedata()

Generate animations from data. This is basically calling plotdata function for multiple snapshots. The main issue here is to determine the colorbar/axis range in advance to avoid any jump in the movie.

VisAna.cutplot — Method

cutplot(data, var; plotrange=[-Inf,Inf,-Inf,Inf], cut=' ',
	plotinterval=0.1, density=1.0, cutPlaneIndex=1,level=20)

2D plane cut contourf of 3D box data.

VisAna.plotdata — Method

plotdata(data, func, (...))

Plot the variable from SWMF output.

plotdata(data, "p", plotmode="contbar")

plotdata(data, "p", plotmode="grid")

plotdata(data, func, plotmode="trimesh",plotrange=plotrange, plotinterval=0.2)

Input arguments

data::Data: output data.

func::String: variables for plotting.

plotmode::String: (optional) type of plotting ["cont","contbar"]...

plotrange::Vector: (optional) range of plotting.

plotinterval: (optional) interval for interpolation.

level: (optional) level of contour.

density: (optional) density for streamlines.

cut: (optional) select 2D cut plane from 3D outputs ["x","y","z"].

cutPlaneIndex: (optional)

multifigure: (optional) 1 for multifigure display, 0 for subplots.

verbose: (optional) display additional information.

Right now this can only deal with 2D plots or 3D cuts. Full 3D plots may be supported in the future.

VisAna.plotlogdata — Method

plotlogdata(data, head, vars, (plotmode="line", plotrange=[-Inf,Inf]))

Plot information from log file.

Input arguments

data::Data: output data.

vars::String: variables for plotting.

plotmode::String: (optional) type of plotting ["line","scatter"].

plotrange::Vector: (optional) range of plotting.

VisAna.select_seeds — Method

select_seeds(x, y, nSeed=100)

Generate nSeed seeding points randomly in the grid range. If you specify nSeed, use the keyword input, otherwise it will be overloaded by the 3D version seed generation.

VisAna.streamslice — Method

streamslice(data::Data, var::String;
  plotrange=[-Inf,Inf,-Inf,Inf], cut=' ',
  plotinterval=0.1, density=1.0, cutPlaneIndex=1,color="w")

Plot streamlines on 2D slices of 3D box data. Variable string must be separated with ;. Tranposes aree needed because of meshgrid and ndgrid conversion.

VisAna.test_dipole — Method

A quick test of the dipole field functions.

VisAna.test_trace_asymptote — Function

testtraceasymptote(IsSingle)

Test streamline tracing by plotting vectors and associated streamlines through a simple velocity field where Vx=x, Vy=-y. Support test for single and double precision.

VisAna.test_trace_dipole — Method

Trace field lines through a dipole field.

VisAna.trace2d_eul — Method

trace2d_eul(fieldx, fieldy, xstart, ystart, gridx, gridy;
	maxstep=20000, ds=0.01, gridType="meshgrid", direction="both")

Given a 2D vector field, trace a streamline from a given point to the edge of the vector field. The field is integrated using Euler's method. While this is faster than rk4, it is less accurate. Only valid for regular grid with coordinates gridx, gridy. If gridx and gridy are not given, assume that xstart and ystart are normalized coordinates (e.g. position in terms of array indices.)??? The field can be in both meshgrid (default) or ndgrid format. Supporting direction for {"both","forward","backward"}.

VisAna.trace2d_rk4 — Method

trace2d_rk4(fieldx, fieldy, xstart, ystart, gridx, gridy;
	maxstep=20000, ds=0.01, gridType="meshgrid", direction="both")

Given a 2D vector field, trace a streamline from a given point to the edge of the vector field. The field is integrated using Runge Kutta 4. Slower than Euler, but more accurate. The higher accuracy allows for larger step sizes ds. Only valid for regular grid with coordinates gridx, gridy. If gridx and gridy are not given, assume that xstart and ystart are normalized coordinates (e.g., position in terms of array indices.???) The field can be in both meshgrid (default) or ndgrid format. Supporting direction for {"both","forward","backward"}.

VisAna.trace3d_eul — Method

trace3d_eul(fieldx, fieldy, fieldz, xstart, ystart, zstart, gridx, gridy,
  gridz; maxstep=20000, ds=0.01)

Given a 3D vector field, trace a streamline from a given point to the edge of the vector field. The field is integrated using Euler's method. Only valid for regular grid with coordinates gridx, gridy, gridz. The field can be in both meshgrid (default) or ndgrid format.