Functions

Functions exported from VisAna:

PyPlot.contour — Function

contour(data, filehead, var, levels=0; plotrange, plotinterval, kwargs)

Wrapper over the contour function in matplotlib.

PyPlot.contourf — Function

contourf(data, filehead, var, levels=0; plotrange, plotinterval, kwargs)

Wrapper over the contourf function in matplotlib.

PyPlot.plot — Method

plot(data, filehead, var; kwargs)

Wrapper over the plot function in matplotlib.

PyPlot.plot_surface — Method

plot_surface(data, filehead, var; plotrange, plotinterval, kwargs)

Wrapper over the plot_surface function in matplotlib.

PyPlot.plot_trisurf — Method

plot_trisurf(data::Data, filehead::Dict, var::String;
	plotrange::Vector{Float64}=[-Inf,Inf,-Inf,Inf], kwargs::Dict=Dict())

Wrapper over the plot_trisurf function in matplotlib.

PyPlot.scatter — Method

scatter(data, filehead, var; kwargs)

Wrapper over the scatter function in matplotlib.

PyPlot.streamplot — Method

streamplot(data, filehead, var; plotrange, plotinterval)

Wrapper over the streamplot function in matplotlib. Streamplot does not have **kwargs in the API.

PyPlot.tricontourf — Method

tricontourf(data, filehead, var; plotrange, plotinterval, kwargs)

Wrapper over the tricontourf function in matplotlib.

VisAna.animatedata — Method

animatedata(filelist, func, (plotmode="contbar",
  plotrange=[-Inf Inf -Inf Inf],
  plotinterval=0.1))

Generate animations from data. This is basically calling plotdata function for multiple snapshots. The main issue here is to determine the colorbar/axis range in advance to avoid any jump in the movie.

VisAna.convertVTK — Function

convertVTK(head, data, connectivity, filename)

Convert 3D unstructured Tecplot data to VTK. Note that if using voxel type data in VTK, the connectivity sequence is different from Tecplot.

VisAna.plotdata — Method

plotdata(data, filehead, func, (...))

Plot the variable from SWMF output.

plotdata(data, filehead, "p", plotmode="contbar")

plotdata(data, filehead, "p", plotmode="grid")

plotdata(data, filehead, func, plotmode="trimesh",plotrange=plotrange, plotinterval=0.2)

Input arguments

data::Data: original variable data.

filehead::Dict: header information.

vars::String: variables for plotting.

plotmode::String: (optional) type of plotting ["cont","contbar"]...

plotrange::Vector: (optional) range of plotting.

plotinterval: (optional) interval for interpolation.

level: (optional) level of contour.

density: (optional) density for streamlines.

cut: (optional) select 2D cut plane from 3D outputs ["x","y","z"].

cutPlaneIndex: (optional)

multifigure: (optional) 1 for multifigure display, 0 for subplots.

verbose: (optional) display additional information.

Right now this can only deal with 2D plots or 3D cuts. Full 3D plots may be supported in the future.

VisAna.plotlogdata — Method

plotlogdata(data, filehead, vars, (plotmode="line", plotrange=[-Inf,Inf]))

Plot information from log file.

Input arguments

data::Data: original variable data.

filehead::Dict: header information.

vars::String: variables for plotting.

plotmode::String: (optional) type of plotting ["line","scatter"].

plotrange::Vector: (optional) range of plotting.

VisAna.readdata — Method

readdata(filenames,(, dir=".", npict=1, verbose=false))

Read data from BATSRUS output files. Stores the npict-th snapshot from an ascii or binary data file into the x [coordinates] and w [data] arrays. Filenames can be provided with wildcards.

Examples

filenames = "1d__raw*"
fileheads, data, filelist = readdata(filenames)

VisAna.readlogdata — Method

readlogdata(filename)

Read information from log file.

VisAna.readtecdata — Function

readtecdata(filename, IsBinary=false, verbose=false)

Return header, data and connectivity from BATSRUS Tecplot outputs. Both binary and ascii formats are supported.

Examples

filename = "3d_ascii.dat"
head, data, connectivity = readtecdata(filename)